+++ package.json
@@ -30,6 +30,7 @@
     "react": "^18.3.1",
     "react-circular-progressbar": "^2.1.0",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.27.0",
     "react-snowfall": "^2.1.1",
     "vite": "^5.4.6",
     "vite-plugin-svgr": "^4.2.0",
